+++ src/api/todoApi.js
@@ -1,4 +1,4 @@
-import axios, { AxiosInstance } from 'axios';
+import axios from 'axios';
 import { GetTokenInStorage } from '../utils/Localstorage';
 
 const BASE_URL = process.env.REACT_APP_API_URL;
@@ -38,25 +38,28 @@ export class todoAPI {
   };
 }
 
-// export const createTodo = async (todo: todo): Promise<TypeTodo> => {
-//   const { data } = await TodoApi.post('/todos', todo);
-//   return data;
-// };
-// export const getTodos = async (): Promise<TypeTodo[]> => {
-//   const { data } = await TodoApi.get('/todos');
-//   return data;
-// };
+const Todoapi = axios.create({
+  baseURL: BASE_URL,
+  headers: {
+    Authorization: `Bearer ${GetTokenInStorage()}`,
+  },
+});
 
-// export const updateTodo = async (
-//   id: updatetodo['id'],
-//   todo: updatetodo['todo'],
-//   isCompleted: updatetodo['isCompleted']
-// ): Promise<TypeTodo> => {
-//   const { data } = await TodoApi.put(`/todos/${id}`, { todo, isCompleted });
-//   return data;
-// };
+export const createTodo = async todo => {
+  const { data } = await Todoapi.post('/todos', todo);
+  return data;
+};
+export const getTodos = async () => {
+  const { data } = await Todoapi.get('/todos');
+  return data;
+};
 
-// export const deleteTodo = async (id: deletetodo['id']): Promise<Status> => {
-//   const data = await TodoApi.delete(`/todos/${id}`);
-//   return data.status;
-// };
+export const updateTodo = async (id, todo, isCompleted) => {
+  const { data } = await Todoapi.put(`/todos/${id}`, { todo, isCompleted });
+  return data;
+};
+
+export const deleteTodo = async id => {
+  const data = await Todoapi.delete(`/todos/${id}`);
+  return data.status;
+};
