+++ src/utils/Localstorage.js
@@ -1,4 +1,4 @@
-const LocalStorageKey = 'Token';
+export const LocalStorageKey = 'Token';
 
 export const SetTokenInStorage = Token => {
   localStorage.setItem(LocalStorageKey, Token);
@@ -11,3 +11,8 @@ export const GetTokenInStorage = () => {
   }
   return false;
 };
+
+// export const isLogin = () => {
+//   console.log(123, !!localStorage.getItem(LocalStorageKey));
+//   return !!localStorage.getItem(LocalStorageKey);
+// };
